### 2020-11-11 #2835 @whedon
[PRE REVIEW]: ADI.jl: A Julia Package for High-Contrast Imaging
**Submitting author:** @mileslucas (<a href="http://orcid.org/0000-0001-6341-310X">Miles Lucas</a>)
**Repository:** <a href="https://github.com/JuliaHCI/ADI.jl" target ="_blank">https://github.com/JuliaHCI/ADI.jl</a>
**Version:** 0.5.2
