### 2020-11-12 #2835 @arfon
:wave: @mileslucas - is this submission associated with a paper in a AAS journal?
