### 2020-11-16 #2835 @dfm
@giordano has agreed to act as our first reviewer - thanks Mosè!

@giordano has disclosed that he and @mileslucas co-mentored a Google Summer of Code project, but he feels that he can provide an impartial review despite this potential conflict of interest. I agree that this potential conflict is minor and believe that the benefit of having Mosè as a reviewer outweighs the apparent conflict of interest. _Therefore, unless @mileslucas has any other concerns, I propose that we declare and waive this conflict_. Miles: please let me know here or via email if you have any questions or concerns!
