### 2020-11-17 #2835 @mileslucas
Thanks Dan (and Mosè)- I have no issues with Mosè reviewing, either, so I'm happy to waive the conflict.

@dfm unrelated to this, a concern about the manuscript is that I currently have a markdown table in it; I'm not sure if that's allowed or not. In its current state the table definitely needs edited to be shown better and I understand pandoc has options for rendering it. However, I understand if the table is a no-go and I can punt it out into my package docs, instead.
