### 2020-11-17 #2835 @dfm
@mileslucas: Sounds good - thanks! For the table, there should be no problem including one, but I agree that it's worth tweaking the formatting. You can use the preview service here to test the compiled version of the manuscript: http://whedon.theoj.org
