### 2020-11-17 #2835 @mileslucas
@dfm if I understand correctly, to tweak the table formatting I would need to change some of the options that pandoc uses when it compiles the pdf. Do I have access to those options in my markdown/bib/meta files?
