### 2020-11-17 #2835 @dfm
Ah! No, I meant that the best approach would be to change the table to fit. I bet that renaming the "pre-processing" column would get you most of the way there. For simplicity, the pandoc options are set in stone!
