### 2020-11-17 #2843 @whedon
[REVIEW]: ADI.jl: A Julia Package for High-Contrast Imaging
**Submitting author:** @mileslucas (<a href="http://orcid.org/0000-0001-6341-310X">Miles Lucas</a>)
**Repository:** <a href="https://github.com/JuliaHCI/ADI.jl" target ="_blank">https://github.com/JuliaHCI/ADI.jl</a>
**Version:** v0.5.3
