### 2020-11-20 JuliaHCI/ADI.jl#26 @giordano
Make `ADIAlgorithm <: Any`?
I might be wrong, but I don't think the fact `ADIAlgorithm <: Function` is used anywhere.  The `Function` type shouldn't really be used to restrict function arguments to functions: there are a lot of "callable" objects which aren't `Function`s, e.g. all type constructors and Python functions from `PyCall`.

I don't have a strong feeling about `ADIAlgorithm <: Function`, but I'd like to hear your opinion about doing simply
```julia
abstract type ADIAlgorithm end
```
