### 2020-11-20 JuliaHCI/ADI.jl#26 @mileslucas
This is a good point. I think this is somewhat of an artifact from my initial design decisions. I don't really see the need to subtype it either, now.
