### 2020-11-20 JuliaHCI/ADI.jl#27 @giordano
`UndefVarError: head not defined` in the documentation
While following the example of the reduction of β Pictoris in the documentation, in the section about the [contrast curve](https://juliahci.github.io/ADI.jl/dev/examples/betapictoris/#Contrast-Curve) I ran into the error
```julia
julia> head(filter(row -> isfinite(row.contrast_corr), cc))
ERROR: UndefVarError: head not defined
Stacktrace:
 [1] top-level scope
   @ REPL[41]:1
```
This is my environment:
```julia
(tmp) pkg> status
Status `/tmp/Project.toml`
  [904a6c7d] ADI v0.5.2
  [3b1b4be9] CFITSIO v1.0.0
  [a93c6f00] DataFrames v0.22.0
  [1893dc6a] HCIDatasets v0.2.0
  [b6cd55e5] HCIToolbox v0.4.2
  [91a5bcdd] Plots v1.9.1
```
