### 2020-11-20 JuliaHCI/ADI.jl#27 @mileslucas
Seems like a compat issue with `DataFrames`. Thanks for pointing it out.
