### 2020-11-20 JuliaHCI/ADI.jl#27 @giordano
Probably related to change of API of `DataFrames`:
```julia
julia> plot(
           cc_empty[:distance],
           [cc_empty[:contrast_corr] cc_empty[:contrast]],
           yscale=:log10,
           xlim=(0, NaN),
           label=["Student-t" "Gaussian"],
           ylabel="5-sigma contrast",
           xlabel="radius [px]"
       )
ERROR: MethodError: no method matching getindex(::DataFrame, ::Symbol)
Closest candidates are:
  getindex(::AbstractDataFrame, ::CartesianIndex{2}) at /home/mose/.julia/packages/DataFrames/X0xNW/src/other/broadcasting.jl:3
  getindex(::AbstractDataFrame, ::Integer, ::Colon) at /home/mose/.julia/packages/DataFrames/X0xNW/src/dataframerow/dataframerow.jl:210
  getindex(::AbstractDataFrame, ::Integer, ::Union{Colon, Regex, AbstractVector{T} where T, All, Between, Cols, InvertedIndex}) at /home/mose/.julia/packages/DataFrames/X0xNW/src/dataframerow/dataframerow.jl:208
  ...
Stacktrace:
 [1] top-level scope
   @ REPL[55]:1
```
I think you need
```julia
cc_empty[:, :distance],
[cc_empty[:, :contrast_corr] cc_empty[:, :contrast]],
```
