### 2020-11-20 2d5f19b
fix dataframes API in examples
