+++ examples/betapictoris.jl
@@ -112,8 +112,8 @@ first(filter(row -> isfinite(row.contrast_corr), cc), 5)
 
 # and lets plot it
 plot(
-    cc[!, :distance],
-    [cc[!, :contrast_corr] cc[!, :contrast]],
+    cc.distance,
+    [cc.contrast_corr cc.contrast],
     yscale=:log10,
     xlim=(0, NaN),
     label=["Student-t" "Gaussian"],
@@ -133,8 +133,8 @@ cc_empty = contrast_curve(alg, cube_empty, angles, psf; fwhm=fwhm, nbranch=6) |>
 first(filter(row -> isfinite(row.contrast_corr), cc_empty), 5)
 #-
 plot(
-    cc_empty[!, :distance],
-    [cc_empty[!, :contrast_corr] cc_empty[!, :contrast]],
+    cc_empty.distance,
+    [cc_empty.contrast_corr cc_empty.contrast],
     yscale=:log10,
     xlim=(0, NaN),
     label=["Student-t" "Gaussian"],
