+++ docs/Project.toml
@@ -11,6 +11,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
 
 [compat]
+DataFrames = "0.22"
 Documenter = "0.25"
 HCIDatasets = "0.2"
 Literate = "2.6"
