### 2020-11-20 b724b00
update subtype of ADIAlgorithm
