### 2020-11-20 JuliaHCI/ADI.jl#26 @mileslucas
Closed via #28
