### 2020-11-26 122c9a3
Fix another typo in beta-pictoris page
