### 2020-11-27 JuliaHCI/ADI.jl#27 @giordano
I can confirm the new version of the documentation fixed the issue, thanks!
