### 2020-11-28 JuliaHCI/ADI.jl#32 @giordano
Error during benchmarks
Since I couldn't install the Python environment on my system (multiple errors for psutil and scipy), I created the following Docker container to reproduce them:
```dockerfile
FROM python:3.8

# Install Julia
RUN curl -q -o - 'https://julialang-s3.julialang.org/bin/linux/x64/1.5/julia-1.5.3-linux-x86_64.tar.gz' | tar xzf -
ENV PATH="/julia-1.5.3/bin:${PATH}"

# Install ADI
RUN julia -e 'import Pkg; Pkg.add(name="ADI", version="0.5.2")'

# Instantiate Julia environment for benchmark
WORKDIR /root/.julia/packages/ADI/OCUJ4/bench
# We need to add ADI because the current Manifest file hard-codes a local path.
RUN julia --project=. -e 'import Pkg; Pkg.add(name="ADI", version="0.5.2"); Pkg.instantiate(); Pkg.precompile()'

# Prepare Python environment
RUN apt-get update -qq
RUN DEBIAN_FRONTEND=noninteractive apt-get install -qq -y gfortran libblas-dev liblapack-dev < /dev/null > /dev/null
# Cleanup apt cache
RUN rm -rf /var/lib/apt/lists/*
RUN pip install -r requirements.txt
```
I had also a problem with the Julia environment, but the solution was to not hard-code the local path of ADI (see #31).

However, when finally running the script `adi.jl` I get the following error:
```
/usr/local/lib/python3.8/site-packages/hciplot/hciplot.py:28: MatplotlibDeprecationWarning: Passing raw data via parameters data and lut to register_cmap() is deprecated since 3.3 and will become an error two minor releases later. Instead use: register_cmap(cmap=LinearSegmentedColormap(name, data, lut))
  register_cmap('ds9cool', data=ds9cool)
/usr/local/lib/python3.8/site-packages/hciplot/hciplot.py:29: MatplotlibDeprecationWarning: Passing raw data via parameters data and lut to register_cmap() is deprecated since 3.3 and will become an error two minor releases later. Instead use: register_cmap(cmap=LinearSegmentedColormap(name, data, lut))
  register_cmap('ds9heat', data=ds9heat)
ERROR: LoadError: KeyError: key :medsub not found
Stacktrace:
 [1] __getproperty(::PyObject, ::Symbol) at /root/.julia/packages/PyCall/BcTLp/src/PyCall.jl:307
 [2] getproperty(::PyObject, ::Symbol) at /root/.julia/packages/PyCall/BcTLp/src/PyCall.jl:312
 [3] top-level scope at /root/.julia/packages/ADI/OCUJ4/bench/adi.jl:33
 [4] include(::String) at ./client.jl:457
 [5] top-level scope at REPL[3]:1
in expression starting at /root/.julia/packages/ADI/OCUJ4/bench/adi.jl:33
```
