### 2020-11-28 0faf89e
update benchmark info in paper
