### 2020-11-28 JuliaHCI/ADI.jl#34 @giordano
Another error duing benchmarks
Following up from #32, this is my new Dockerfile:
```dockerfile
FROM python:3.8

# Install Julia
RUN curl -q -o - 'https://julialang-s3.julialang.org/bin/linux/x64/1.5/julia-1.5.3-linux-x86_64.tar.gz' | tar xzf -
ENV PATH="/julia-1.5.3/bin:${PATH}"

# Install ADI
RUN julia -e 'using Pkg; pkg"dev ADI"'

# Instantiate Julia environment for benchmark
WORKDIR /root/.julia/dev/ADI/bench
# We need to dev ADI because the current Manifest file hard-codes a local path.
RUN julia --project=. -e 'using Pkg; pkg"dev ADI"; Pkg.instantiate(); Pkg.precompile()'

# Prepare Python environment
RUN apt-get update -qq
RUN DEBIAN_FRONTEND=noninteractive apt-get install -qq -y gfortran libblas-dev liblapack-dev < /dev/null > /dev/null
# Cleanup apt cache
RUN rm -rf /var/lib/apt/lists/*
RUN pip install -r requirements.txt

# Prefetch HCI data sets
ENV DATADEPS_ALWAYS_ACCEPT=true
RUN julia --project=. -e 'using HCIDatasets: BetaPictoris, HR8799; BetaPictoris[:cube]; BetaPictoris[:pa]; HR8799[:cube]; HR8799[:pa]'
```
Now when running `bench/adi.jl` I get:
```
[ Info: Benchmarking - HCIDatasets.BetaPictorisDataset() ADI Reduction
[ Info: Python - #6
ERROR: LoadError: PyError ($(Expr(:escape, :(ccall(#= /root/.julia/packages/PyCall/BcTLp/src/pyfncall.jl:43 =# @pysym(:PyObject_Call), PyPtr, (PyPtr, PyPtr, PyPtr), o, pyargsptr, kw))))) <class 'RuntimeError'>
RuntimeError('Opencv python bindings cannot be imported. Install opencv or set imlib to skimage')
  File "/usr/local/lib/python3.8/site-packages/vip_hci/medsub/medsub_source.py", line 168, in median_sub
    cube_der = cube_derotate(cube_out, angle_list, imlib=imlib,
  File "/usr/local/lib/python3.8/site-packages/vip_hci/preproc/derotation.py", line 199, in cube_derotate
    array_der[i] = frame_rotate(array[i], -angle_list[i], imlib=imlib,
  File "/usr/local/lib/python3.8/site-packages/vip_hci/preproc/derotation.py", line 118, in frame_rotate
    raise RuntimeError(msg)

Stacktrace:
 [1] pyerr_check at /root/.julia/packages/PyCall/BcTLp/src/exception.jl:62 [inlined]
 [2] pyerr_check at /root/.julia/packages/PyCall/BcTLp/src/exception.jl:66 [inlined]
 [3] _handle_error(::String) at /root/.julia/packages/PyCall/BcTLp/src/exception.jl:83
 [4] macro expansion at /root/.julia/packages/PyCall/BcTLp/src/exception.jl:97 [inlined]
 [5] #110 at /root/.julia/packages/PyCall/BcTLp/src/pyfncall.jl:43 [inlined]
 [6] disable_sigint at ./c.jl:446 [inlined]
 [7] __pycall! at /root/.julia/packages/PyCall/BcTLp/src/pyfncall.jl:42 [inlined]
 [8] _pycall!(::PyObject, ::PyObject, ::Tuple{Array{Float32,3},Array{Float32,1}}, ::Int64, ::PyObject) at /root/.julia/packages/PyCall/BcTLp/src/pyfncall.jl:29
 [9] _pycall!(::PyObject, ::PyObject, ::Tuple{Array{Float32,3},Array{Float32,1}}, ::Base.Iterators.Pairs{Symbol,Bool,Tuple{Symbol},NamedTuple{(:verbose,),Tuple{Bool}}}) at /root/.julia/packages/PyCall/BcTLp/src/pyfncall.jl:11
 [10] #_#117 at /root/.julia/packages/PyCall/BcTLp/src/pyfncall.jl:86 [inlined]
 [11] (::var"#6#12")(::Array{Float32,3}, ::Vararg{Any,N} where N; kwargs::Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}}) at /root/.julia/dev/ADI/bench/adi.jl:34
 [12] #6 at /root/.julia/dev/ADI/bench/adi.jl:34 [inlined]
 [13] ##core#295(::var"#6#12", ::Array{Float32,3}, ::Array{Float32,1}) at /root/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:371
 [14] ##sample#296(::BenchmarkTools.Parameters) at /root/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:377
 [15] _run(::BenchmarkTools.Benchmark{Symbol("##benchmark#294")}, ::BenchmarkTools.Parameters; verbose::Bool, pad::String, kwargs::Base.Iterators.Pairs{Symbol,Integer,NTuple{4,Symbol},NamedTuple{(:samples, :evals, :gctrial, :gcsample),Tuple{Int64,Int64,Bool,Bool}}}) at /root/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:405
 [16] (::Base.var"#inner#2"{Base.Iterators.Pairs{Symbol,Integer,NTuple{5,Symbol},NamedTuple{(:verbose, :samples, :evals, :gctrial, :gcsample),Tuple{Bool,Int64,Int64,Bool,Bool}}},typeof(BenchmarkTools._run),Tuple{BenchmarkTools.Benchmark{Symbol("##benchmark#294")},BenchmarkTools.Parameters}})() at ./essentials.jl:713
 [17] #invokelatest#1 at ./essentials.jl:714 [inlined]
 [18] #run_result#37 at /root/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:32 [inlined]
 [19] run(::BenchmarkTools.Benchmark{Symbol("##benchmark#294")}, ::BenchmarkTools.Parameters; progressid::Nothing, nleaves::Float64, ndone::Float64, kwargs::Base.Iterators.Pairs{Symbol,Integer,NTuple{5,Symbol},NamedTuple{(:verbose, :samples, :evals, :gctrial, :gcsample),Tuple{Bool,Int64,Int64,Bool,Bool}}}) at /root/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:94
 [20] #warmup#45 at /root/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:141 [inlined]
 [21] warmup(::BenchmarkTools.Benchmark{Symbol("##benchmark#294")}) at /root/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:141
 [22] top-level scope at /root/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:287
 [23] top-level scope at /root/.julia/dev/ADI/bench/adi.jl:47
 [24] include(::String) at ./client.jl:457
 [25] top-level scope at REPL[1]:1
in expression starting at /root/.julia/dev/ADI/bench/adi.jl:39
```
Not sure what I'm supposed to do, whether to install opencv or set imlib to skimage.
