### 2020-11-29 6805605
CompatHelper: bump compat for "StaticArrays" to "1.0"
