+++ Project.toml
@@ -37,7 +37,7 @@ ProgressLogging = "0.1"
 Reexport = "0.2"
 Rotations = "1"
 Setfield = "0.7"
-StaticArrays = "0.12"
+StaticArrays = "0.12, 1.0"
 StaticKernels = "0.5"
 StatsBase = "0.33"
 TSVD = "0.4"
