### 2020-12-01 JuliaHCI/ADI.jl#34 @mileslucas
So I took a few shots at getting a working dockerfile, but I never got the opencv bindings to import correctly through VIP.

You could chooose to run the benchmarks against skimage, by changing https://github.com/JuliaHCI/ADI.jl/blob/570fede975d3cd329775a7c06657a751d5cc70fa/bench/adi.jl#L47
to
```julia
@belapsed $alg($cube, $angles; imlib="skimage")
```
however, the VIP documentation notes opencv is faster than skimage, so the benchmarks may be discrepant.

I forget what the initial issue was on your local machine, but I'm quickly running out of ideas since the docker builds aren't working.
