### 2020-12-01 JuliaHCI/ADI.jl#34 @giordano
> I forget what the initial issue was on your local machine, but I'm quickly running out of ideas since the docker builds aren't working.

Deoending on whether I was using system's pip or conda I was unable to install psutil or scipy. Using a Docker container looked like a good idea to try and reproduce the benchmarks in a consistent way, but instantiating the correct python environment seems challenging
