### 2020-12-01 JuliaHCI/ADI.jl#34 @giordano
I'm a bit lost though: the README of VIP [claims](https://github.com/vortex-exoplanet/VIP/blob/df97f3037f6f43dfa13d7d0b12897d9f014517a9/README.rst#other-dependencies)

> ``OpenCV`` (Open source Computer Vision) provides fast C++ image processing
> operations and is used by ``VIP`` for basic image transformations. If you don't
> have/want the ``OpenCV`` python bindings (``OpenCV`` is optional since ``VIP``
> v0.5.2), ``VIP`` will use the much slower ``ndimage``/``scikit-image`` libraries
> transparently. Fortunately, installing ``OpenCV`` library is nowadays and easy
> process that is done automatically with the ``VIP`` installation.

Why doesn't that work? :confused:
