### 2020-12-01 JuliaHCI/ADI.jl#34 @giordano
```
root@702c7b1535ca:~/.julia/dev/ADI/bench# ldd /usr/local/lib/python3.8/site-packages/cv2/cv2.cpython-38-x86_64-linux-gnu.so
        linux-vdso.so.1 (0x00007ffe968ea000)
        libpng15-c2ffaf3d.so.15.13.0 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libpng15-c2ffaf3d.so.15.13.0 (0x00007f39e6f4b000)
        libavcodec-8daa01ff.so.58.109.100 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libavcodec-8daa01ff.so.58.109.100 (0x00007f39e5b31000)
        libavformat-06a336f2.so.58.61.100 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libavformat-06a336f2.so.58.61.100 (0x00007f39e58b3000)
        libavutil-01d48d95.so.56.60.100 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libavutil-01d48d95.so.56.60.100 (0x00007f39e55f4000)
        libswscale-2d2bce5d.so.5.8.100 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libswscale-2d2bce5d.so.5.8.100 (0x00007f39e5560000)
        libQt5Widgets-0e1d98b5.so.5.15.0 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libQt5Widgets-0e1d98b5.so.5.15.0 (0x00007f39e4c1d000)
        libQt5Gui-ca084835.so.5.15.0 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libQt5Gui-ca084835.so.5.15.0 (0x00007f39e4286000)
        libQt5Test-9d4ee57e.so.5.15.0 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libQt5Test-9d4ee57e.so.5.15.0 (0x00007f39e4216000)
        libQt5Concurrent-21c64331.so.5.15.0 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libQt5Concurrent-21c64331.so.5.15.0 (0x00007f39e4207000)
        libQt5Core-d92f2856.so.5.15.0 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libQt5Core-d92f2856.so.5.15.0 (0x00007f39e3aaf000)
        libz-d8a329de.so.1.2.7 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libz-d8a329de.so.1.2.7 (0x00007f39e3896000)
        libdl.so.2 => /lib/x86_64-linux-gnu/libdl.so.2 (0x00007f39e3888000)
        libpthread.so.0 => /lib/x86_64-linux-gnu/libpthread.so.0 (0x00007f39e3867000)
        librt.so.1 => /lib/x86_64-linux-gnu/librt.so.1 (0x00007f39e385d000)
        libstdc++.so.6 => /usr/lib/x86_64-linux-gnu/libstdc++.so.6 (0x00007f39e36d9000)
        libm.so.6 => /lib/x86_64-linux-gnu/libm.so.6 (0x00007f39e3556000)
        libgcc_s.so.1 => /lib/x86_64-linux-gnu/libgcc_s.so.1 (0x00007f39e353a000)
        libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007f39e3379000)
        /lib64/ld-linux-x86-64.so.2 (0x00007f39eaba4000)
        libswresample-4767dc06.so.3.8.100 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libswresample-4767dc06.so.3.8.100 (0x00007f39e3358000)
        libvpx-14094576.so.6.3.0 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libvpx-14094576.so.6.3.0 (0x00007f39e3042000)
        libbz2-a273e504.so.1.0.6 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libbz2-a273e504.so.1.0.6 (0x00007f39e2e31000)
        libssl-f3db6a3b.so.1.1 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libssl-f3db6a3b.so.1.1 (0x00007f39e2d8c000)
        libcrypto-098682aa.so.1.1 => /usr/local/lib/python3.8/site-packages/cv2/../opencv_python.libs/libcrypto-098682aa.so.1.1 (0x00007f39e2a75000)
        libGL.so.1 => not found
        libGL.so.1 => not found
        libgthread-2.0.so.0 => /usr/lib/x86_64-linux-gnu/libgthread-2.0.so.0 (0x00007f39e2a70000)
        libglib-2.0.so.0 => /usr/lib/x86_64-linux-gnu/libglib-2.0.so.0 (0x00007f39e294f000)
        libpcre.so.3 => /lib/x86_64-linux-gnu/libpcre.so.3 (0x00007f39e28db000)
```
Uh oh, it looks like OpenGL is missing
