### 2020-12-01 JuliaHCI/ADI.jl#34 @giordano
Ok, I confirm that I had also to install `mesa-utils` in order to be able to run the benchmarks.  This is my final Dockerfile:
```dockerfile
FROM python:3.8

# Install Julia
RUN curl -q -o - 'https://julialang-s3.julialang.org/bin/linux/x64/1.5/julia-1.5.3-linux-x86_64.tar.gz' | tar xzf -
ENV PATH="/julia-1.5.3/bin:${PATH}"

# Install ADI
RUN julia -e 'using Pkg; pkg"dev ADI"'

# Instantiate Julia environment for benchmark
WORKDIR /root/.julia/dev/ADI/bench
# We need to dev ADI because the current Manifest file hard-codes a local path.
RUN julia --project=. -e 'using Pkg; pkg"dev ADI"; Pkg.instantiate(); Pkg.precompile()'
# Prefetch HCI data sets
ENV DATADEPS_ALWAYS_ACCEPT=true
RUN julia --project=. -e 'using HCIDatasets: BetaPictoris, HR8799; BetaPictoris[:cube]; BetaPictoris[:pa]; HR8799[:cube]; HR8799[:pa]'

# Prepare Python environment
RUN apt-get update -qq
# Install `gfortran`, `libblas-dev`, `liblapack-dev` to build scipy,
# `mesa-utils` is needed by OpenCV.
RUN DEBIAN_FRONTEND=noninteractive apt-get install -qq -y gfortran libblas-dev liblapack-dev mesa-utils < /dev/null > /dev/null
# Cleanup apt cache
RUN rm -rf /var/lib/apt/lists/*
RUN pip install -r requirements.txt
```
