### 2020-12-01 JuliaHCI/ADI.jl#34 @mileslucas
Nice, I'm glad you managed to find a solution! This Dockerfile will be nice for posterity, too.
