### 2020-12-02 962b840
add import info to installation section
