### 2020-12-02 6d08692
add more info about formulae to getting started
