### 2020-12-02 cbc4833
add notes about HCItoolbox reexportation
