### 2020-12-02 a6b712c
fix some typos
