### 2020-12-03 6e40f2b
CompatHelper: bump compat for "StaticKernels" to "0.6"
