+++ Project.toml
@@ -38,7 +38,7 @@ Reexport = "0.2"
 Rotations = "1"
 Setfield = "0.7"
 StaticArrays = "0.12, 1.0"
-StaticKernels = "0.5"
+StaticKernels = "0.5, 0.6"
 StatsBase = "0.33"
 TSVD = "0.4"
 UnPack = "1"
