### 2020-12-03 25aeb77
add stim map
