### 2020-12-03 6364745
add tests for stim
