+++ paper/paper.md
@@ -30,7 +30,7 @@ The difference in ADI algorithms is primarily in how the systematic signal is es
 
 # Statement of need
 
-`ADI.jl` is a Julia framework for post-processing high-contrast imaging (HCI) data. By organizing algorithms separately from their application, `ADI.jl` offers a modular API the benefits both observers and algorithm designers. Observers can rapidly experiment with multiple post-processing algorithms and techniques to optimally reduce their data. Julia's dynamic just-in-time (JIT) LLVM compiler [@Julia-2017; @2018arXiv180803370B] means this experimentation comes at a low runtime cost to the observer, enabling broader experimentation or higher throughput, for example, in the case of survey pipelines.
+`ADI.jl` is a Julia framework for post-processing high-contrast imaging (HCI) data. By organizing algorithms separately from their application, `ADI.jl` offers a modular API that benefits both observers and algorithm designers. Observers can rapidly experiment with multiple post-processing algorithms and techniques to optimally reduce their data. Julia's dynamic just-in-time (JIT) LLVM compiler [@Julia-2017; @2018arXiv180803370B] means this experimentation comes at a low runtime cost to the observer, enabling broader experimentation or higher throughput, for example, in the case of survey pipelines.
 
 Algorithm designers will find that Julia is highly composable, so extending or adding a new algorithm only requires writing the code that is *unique to that algorithm*. Julia's language interoperability also means the algorithm can be implemented in Python or C, for example. In other words, to be able to fully use the post-processing capabilities of `ADI.jl` a new algorithm only needs to implement one or two methods. Furthermore, computational techniques like GPU programming are available *generically* through packages like `CUDA.jl` [@besard2018juliagpu].
 
