### 2020-12-03 4f90ff2
update docs about sdi and data formats
