### 2020-12-03 a8086bf
some paper revisions
