### 2020-12-04 534f159
change stim <--> stimmap and clean up some docstrings
