+++ src/metrics/contrast.jl
@@ -8,27 +8,17 @@ using StaticKernels
 using LinearAlgebra: dot
 
 """
-    contrast_curve(alg,
-                   cube,
-                   angles,
-                   psf,
-                   args...;
-                   fwhm,
-                   sigma=5,
+    contrast_curve(alg, cube, angles, psf, args...;
+                   fwhm, sigma=5, nbranch=1, theta=0, inner_rad=1,
                    starphot=Metrics.estimate_starphot(cube, fwhm),
-                   nbranch=1,
-                   theta=0,
-                   inner_rad=1,
-                   fc_rad_sep=3,
-                   snr=100,
-                   k=2,
-                   kwargs...)
+                   fc_rad_sep=3, snr=100, k=2, smooth=true,
+                   subsample=true, kwargs...)
 
 Calculate the throughput-calibrated contrast. This first processes the algorithmic [`throughput`](@ref) by injecting instances of `psf` into `cube`. These are processed through `alg` and the ratio of the recovered flux to the injected flux is calculated. These companions are injected in resolution elements across the frame, which can be changed via the various keyword arguments.
 
 The throughput can only be calculated for discrete resolution elements, but we typically want a much smoother curve. To accomplish this, we measure the noise (the standard deviation of all resolution elements in an annulus at a given radius) for every pixel in increasing radii. We then interpolate the throughput to this grid and return the subsampled curves.
 
-# Fields
+# Returned Fields
 * `distance` - The radial distance (in pixels) for each measurement
 * `contrast` - The Gaussian sensitivity
 * `contrast_corr` - The Student-t sensitivity
@@ -55,56 +45,27 @@ The throughput can only be calculated for discrete resolution elements, but we t
     contrast_curve(alg, cube, angles, psf; fwhm=fwhm) |> DataFrame
     ```
 """
-function contrast_curve(alg,
-                        cube,
-                        angles,
-                        psf,
-                        args...;
-                        fwhm,
-                        starphot=estimate_starphot(cube, fwhm),
-                        sigma=5,
-                        inner_rad=1,
-                        fc_rad_sep=3,
-                        theta=0,
-                        nbranch=1,
-                        snr=100,
-                        subsample=true,
-                        k=2,
-                        smooth=true,
-                        kwargs...)
+function contrast_curve(alg, cube, angles, psf, args...;
+        fwhm, sigma=5, nbranch=1, theta=0, inner_rad=1,
+        starphot=Metrics.estimate_starphot(cube, fwhm),
+        fc_rad_sep=3, snr=100, k=2, smooth=true,
+        subsample=true, kwargs...)
 
     # measure the noise and throughput in consecutive resolution elements
     # across azimuthal branches
     @info "Calculating Throughput"
     reduced_empty = alg(cube, angles, args...; kwargs...)
 
-    through, meta = throughput(alg,
-                               cube,
-                               angles,
-                               psf,
-                               args...;
-                               fwhm=fwhm,
-                               inner_rad=inner_rad,
-                               fc_rad_sep=fc_rad_sep,
-                               theta=theta,
-                               nbranch=nbranch,
-                               snr=snr,
-                               reduced_empty=reduced_empty,
-                               kwargs...)
+    through, meta = throughput(alg, cube, angles, psf, args...;
+                               fwhm=fwhm, inner_rad=inner_rad, fc_rad_sep=fc_rad_sep, theta=theta,
+                               nbranch=nbranch, snr=snr, reduced_empty=reduced_empty, kwargs...)
 
[…]
