+++ src/metrics/stim.jl
@@ -1,7 +1,7 @@
 using Statistics
 
 """
-    stim(residuals, angles)
+    stimmap(residuals, angles)
 
 Calculate the standardized trajectory intensity mean (STIM) map. The inputs are a cube of residuals and the corresponding parallactic angles.
 
@@ -9,7 +9,7 @@ This method seeks to improve upon the typical student-t S/N tests ([`snr`](@ref)
 
 In particular, the STIM map is robust to detections with multiple objects or extended sources within the same annuli, which results in very high noise estimates using spatial methods. The STIM map also performs better at small angular separations, since the temporal domain has no limitations from limited resolution elements.
 
-*Pairet et al. 2019* derives a detection threshold of `τ ≈ 0.5` for the STIM map. The detection threshold can be calculated in a similar manner using [`stim_threshold`](@ref).
+*Pairet et al. 2019* derives a detection threshold of `τ ≈ 0.5` for the STIM map. The detection threshold can be calculated for a specific dataset using [`stim_threshold`](@ref).
 
 # Examples
 
@@ -20,7 +20,7 @@ julia> L = reconstruct(PCA(10), cube, angles);
 
 julia> S = cube .- L;
 
-julia> stimmap = stim(S, angles);
+julia> sm = stimmap(S, angles);
 ```
 
 # References
@@ -29,29 +29,44 @@ julia> stimmap = stim(S, angles);
 # See Also
 [`stim_threshold`](@ref)
 """
-function stim(residuals::AbstractArray{T,3}, angles) where T
-    return stimmap(derotate(residuals, angles))
+function stimmap(residuals::AbstractArray{T,3}, angles) where T
+    return stim(derotate(residuals, angles))
 end
 
 """
     stim_threshold([stimmap, ] residuals, angles)
 
-Calculate the detection threshold for the standardized trajectory intensity mean (STIM) map. This method uses the same residual cube as [`stim`](@ref) but adds an additional step of estimating the residual noise by derotating the residuals with the *opposite* parallactic angles.
+Calculate the detection threshold for the standardized trajectory intensity mean (STIM) map. This method uses the same residual cube as [`stimmap`](@ref) but adds an additional step of estimating the residual noise by derotating the residuals with the *opposite* parallactic angles.
 
 If the STIM map has already been calculated, it can be passed in, otherwise it will be calculated in addition to the noise map. Note this will not return the STIM map, only the threshold.
 
 The threshold is derived in section 5.1 of *Pairet et al. 2019* as the ratio of the number of pixels above the approximated noise map.
 
+# Examples
+
+
+```julia
+julia> cube, angles = # load data
+
+julia> L = reconstruct(PCA(10), cube, angles);
+
+julia> S = cube .- L;
+
+julia> sm = stimmap(S, angles);
+
+julia> τ = stim_threshold(sm, S, angles);
+```
+
 # References
 * [Pairet et al. 2019](http://adsabs.harvard.edu/abs/2019MNRAS.487.2262P) "STIM map: detection map for exoplanets imaging beyond asymptotic Gaussian residual speckle noise"
 
 # See Also
-[`stim`](@ref)
+[`stimmap`](@ref)
 """
 function stim_threshold(stimmap, residuals, angles)
     # estimate noise map by getting STIM map of cube
     # rotated with opposite angles
-    d_opp = Metrics.stimmap(derotate(residuals, -angles))
+    d_opp = Metrics.stim(derotate(residuals, -angles))
     # return ratio of values above the noise
     n_ϵ = count(stimmap .> d_opp)
     n = length(stimmap)
@@ -60,16 +75,16 @@ end
 
[…]
