+++ docs/src/metrics.md
@@ -1,6 +1,35 @@
 # Metrics
 
-## API/Reference
-```@autodocs
+```@index
 Modules = [ADI.Metrics]
+```
+
+```@docs
+ADI.Metrics
+```
+
+## Detection Maps
+
+```@docs
+detectionmap
+snr
+significance
+noise
+stimmap
+stim_threshold
+Metrics.stim
+```
+
+## Throughput
+
+```@docs
+throughput
+```
+
+## Contrast Curve
+
+```@docs
+contrast_curve
+Metrics.subsample_contrast
+Metrics.estimate_starphot
 ```
