### 2020-12-04 1e4cfd7
add community guidelines and support info
