+++ docs/src/index.md
@@ -31,6 +31,13 @@ julia> using ADI
 
 For more information, see the [Pkg documentation](https://docs.julialang.org/en/v1/stdlib/Pkg/).
 
+## Contributing and Support
+
+[![ColPrac: Contributor's Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-9cf)](https://github.com/SciML/ColPrac)
+
+In general, all contributions should follow [ColPrac](https://github.com/SciML/ColPrac). Feel free to open an issue or reach out to the developers to coordinate a contribution or discuss ideas. For support with using ADI.jl, please open an [issue](
+https://github.com/JuliaHCI/ADI.jl/issues/new/) describing the problem and steps to reproduce it.
+
 ## License
 
 This work is distributed under the MIT "expat" license. See [`LICENSE`](https://github.com/juliahci/ADI.jl/blob/master/LICENSE) for more information.
