+++ README.md
@@ -74,6 +74,13 @@ julia> df = DataFrame(cc)
 julia> head(df)
 ```
 
+## Contributing and Support
+
+[![ColPrac: Contributor's Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-9cf)](https://github.com/SciML/ColPrac)
+
+In general, all contributions should follow [ColPrac](https://github.com/SciML/ColPrac). Feel free to open an issue or reach out to the developers to coordinate a contribution or discuss ideas. For support with using ADI.jl, please open an [issue](
+https://github.com/JuliaHCI/ADI.jl/issues/new/) describing the problem and steps to reproduce it.
+
 ## License
 
 This package is licensed under the MIT Expat license. See [LICENSE](LICENSE) for more information.
