### 2020-12-04 d399b59
tidy up comparison table
