+++ docs/src/index.md
@@ -33,10 +33,9 @@ For more information, see the [Pkg documentation](https://docs.julialang.org/en/
 
 ## Contributing and Support
 
-[![ColPrac: Contributor's Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-9cf)](https://github.com/SciML/ColPrac)
+[![ColPrac: Contributor's Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-blueviolet)](https://github.com/SciML/ColPrac)
 
-In general, all contributions should follow [ColPrac](https://github.com/SciML/ColPrac). Feel free to open an issue or reach out to the developers to coordinate a contribution or discuss ideas. For support with using ADI.jl, please open an [issue](
-https://github.com/JuliaHCI/ADI.jl/issues/new/) describing the problem and steps to reproduce it.
+In general, all contributions should follow [ColPrac](https://github.com/SciML/ColPrac). Feel free to open an issue or reach out to the developers to coordinate a contribution or discuss ideas. For support with using ADI.jl, please open an [issue](https://github.com/juliahci/ADI.jl/issues/new/) describing the problem and steps to reproduce it.
 
 ## License
 
