### 2020-12-04 7c89919
remove table from paper
