+++ paper/paper.md
@@ -40,18 +40,9 @@ Currently `ADI.jl` supports full-frame ADI and RDI processing, with experimental
 
 High-contrast imaging as a field predominantly utilizes Python for data reduction. We break down some of the necessary computations into *pre-processing*, which includes raw calibration of data, the centering and stacking of the data cube, bad-pixel removal, etc., *post-processing*, which includes the PSF approximation and subtraction, *detection metrics* which includes methods for analyzing post-processed data to make detections or find limits of detections, and finally *forward modeling* which includes various statistical models for companions and disks that can be used with post-processing algorithms in a maximum likelihood framework. `ADI.jl` primarily focuses on post-processing and detection metrics.
 
-Some notable libraries for HCI tasks include the Vortex Imaging Pipeline (`VIP`) [@gomez_gonzalez_vip_2016], `pyKLIP` [@2015ascl.soft06001W], and `PynPoint` [@pynpoint:2019]. A table of the feature sets of these packages alongside `ADI.jl` is presented below. In particular, `VIP` has served as a useful source of information regarding HCI image-processing as well as detailed implementations of common ADI algorithms. This has been indispensable in the development of `ADI.jl`, although this package is not a direct translation. In our small benchmark suite, `ADI.jl` is roughly the same speed or slightly quicker than `VIP` for the algorithms we tested, except for NMF, while detection maps were ~2 orders of magnitude quicker.
+Some notable libraries for HCI tasks include the Vortex Imaging Pipeline (`VIP`) [@gomez_gonzalez_vip_2016], `pyKLIP` [@2015ascl.soft06001W], and `PynPoint` [@pynpoint:2019]. A table of the feature sets of these packages alongside `ADI.jl` is presented in the [online documentation](https://juliahci.github.io/ADI.jl/dev/gettingstarted/#Feature-comparison). In particular, `VIP` has served as a useful source of information regarding HCI image-processing as well as detailed implementations of common ADI algorithms. This has been indispensable in the development of `ADI.jl`, although this package is not a direct translation. In our small benchmark suite, `ADI.jl` is roughly the same speed or slightly quicker than `VIP` for the algorithms we tested, except for NMF, while detection maps were ~2 orders of magnitude quicker.
 
 In general `VIP` offers the most diversity in algorithms and their applications, but not all algorithms are as feature-complete as the PCA implementation. `VIP` also contains many useful utilities for pre-processing and a pipeline framework. `pyKLIP` primarily uses the PCA (KLIP) algorithm, but offers many forward modeling implementations. `PynPoint` has a highly modular pre-processing module that is focused on pipelines.
-
-Table 1: Comparison of features across different HCI frameworks. Techniques marked with * indicate partial support, meaning that not all algorithms are supported.
-
-| Framework | Pre-processing | Algorithms | Techniques | Detection Metrics | Forward Modeling |
-|:-:|:-:|:-|:-|:-|:-|
-| ADI.jl   | no   | median, PCA, NMF, fixed-point GreeDS   | Full-frame ADI/RDI, SDI (experimental)   | detection maps, STIM, contrast curve, STIM   | no   |
-| VIP  | yes  | median, LOCI, PCA, NMF, LLSG, ANDROMEDA, pairwise frame differencing  | Full-frame ADI/RDI, SDI, annular ADI/RDI*  | detection maps, blob detection, STIM, ROC, contrast curve  | NegFC  |
[…]
