### 2020-12-04 JuliaHCI/ADI.jl#30 @mileslucas
Thanks for all the in-depth comments and suggestions, I really appreciate them. I've made a slew of documentation revisions; thanks for pointing out locations lacking in context, especially.

Your point 5 may be due to mismatched docs and code- I didn't start re-exporting HCIToolbox until version `0.5.3` (which is the master branch), so if you want to try again I'd recommend doing a fresh pull.

For point 7 I've added contribution and support section to the README and homepage of the docs.

Again, thanks for all the help!
