### 2020-12-06 d429c63
add multi-threading note to benchmarks
