+++ Project.toml
@@ -4,6 +4,7 @@ authors = ["Miles Lucas <mdlucas@hawaii.edu>"]
 version = "0.5.3"
 
 [deps]
+CoordinateTransformations = "150eb455-5306-5404-9cee-2592286d6298"
 Dierckx = "39dd38d3-220a-591b-8e3c-4c3a8c710a94"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
@@ -25,6 +26,7 @@ TSVD = "9449cd9e-2762-5aa3-a617-5413e99d722e"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [compat]
+CoordinateTransformations = "0.5, 0.6"
 Dierckx = "0.4, 0.5"
 Distributions = "0.23, 0.24"
 FillArrays = "0.8, 0.9, 0.10"
