### 2020-12-11 a6d3781
add single-spot throughput method and notes about photometry
