+++ src/median.jl
@@ -6,7 +6,7 @@ using HCIToolbox: get_annulus_segments
 Classic PSF subtraction using the median of entire data cube.
 
 # References
-* [Marois et al. 2006](http://adsabs.harvard.edu/abs/2006ApJ...641..556M) Angular Differential Imaging: A Powerful High-Contrast Imaging Technique
+1. [Marois et al. 2006](http://adsabs.harvard.edu/abs/2006ApJ...641..556M) Angular Differential Imaging: A Powerful High-Contrast Imaging Technique
 """
 struct Median <: ADIAlgorithm end
 
