### 2020-12-11 dcaab6f
use more accurate fwhm in example
