### 2020-12-11 1e53c39
update NMF docs and implementation to warn on negative inputs
