### 2020-12-13 #2843 @giordano
I also just finished the checklist.  I'd like to mention that I had left some comments in `ADI.jl` bug tracker:

* https://github.com/JuliaHCI/ADI.jl/issues/26 about subtyping of a data structure
* https://github.com/JuliaHCI/ADI.jl/issues/27 about an error experienced while going through examples in the documentation
* https://github.com/JuliaHCI/ADI.jl/issues/30 more general suggestions about the documentation
* https://github.com/JuliaHCI/ADI.jl/issues/32 and https://github.com/JuliaHCI/ADI.jl/issues/34 reporting errors while running benchmarks

All reported issues have been resolved already.  I also opened a couple of pull requests with trivial changes:

* https://github.com/JuliaHCI/ADI.jl/pull/29 to fix a typo in documentation
* https://github.com/JuliaHCI/ADI.jl/pull/31 to make it possible to load the Julia environment for benchmarks
