### 2020-12-13 #2843 @mileslucas
I'll mention, too, I have the working changes on the master branch ready to release, which will change the version to `0.5.3` (rather than the `0.5.2` when I initiated the submission).
