### 2020-12-15 2fc15c9
Bibliography edits
