+++ paper/paper.bib
@@ -198,7 +198,7 @@ archivePrefix = {arXiv},
        author = {{Stolker}, T. and {Bonse}, M.~J. and {Quanz}, S.~P. and {Amara}, A. and
          {Cugno}, G. and {Bohn}, A.~J. and {Boehle}, A.},
         title = "{PynPoint: a modular pipeline architecture for processing and analysis of high-contrast imaging data}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {methods: data analysis, techniques: high angular resolution, techniques: image processing, planets and satellites: detection, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2019,
         month = jan,
@@ -213,7 +213,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@MISC{2015ascl.soft06001W,
+@article{2015ascl.soft06001W,
        author = {{Wang}, Jason J. and {Ruffio}, Jean-Baptise and {De Rosa}, Robert J. and
          {Aguilar}, Jonathan and {Wolff}, Schuyler G. and {Pueyo}, Laurent},
         title = "{pyKLIP: PSF Subtraction for Exoplanets and Disks}",
@@ -224,6 +224,7 @@ archivePrefix = {arXiv},
         pages = {ascl:1506.001},
 archivePrefix = {ascl},
        eprint = {1506.001},
+      journal = {Astrophysics Source Code Library},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2015ascl.soft06001W},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
