### 2020-12-15 b3a5aa7
Small manuscript edits
