+++ paper/paper.md
@@ -22,11 +22,11 @@ bibliography: paper.bib
 
 # Summary
 
-High-contrast imaging (HCI) is a powerful technique for discovering and characterizing exoplanets. Being able to probe the architecture, formation, and atmospheres of planets *directly* is necessary for advancing companion formation and evolution theory [@bowler_imaging_2016]. The process required to image an exoplanet is daunting, however, due to the brightness and proximity of their host stars. One part of making such a difficult detection is the image processing of HCI data to remove systematic signal and attenuate noise. The size of HCI data and complexity of processing algorithms requires an efficient numerical framework that simultaneously offers modularity for rapid experimentation and discovery.
+High-contrast imaging (HCI) is a powerful technique for discovering and characterizing exoplanets. Being able to probe the architecture, formation, and atmospheres of planets *directly* is necessary for advancing companion formation and evolution theory [@bowler_imaging_2016]. The process required to image an exoplanet is daunting, however, due to the brightness and proximity of their host stars. One part of making such a difficult detection is the image processing of HCI data to remove systematic signals and attenuate noise. The size of HCI data and complexity of processing algorithms requires an efficient numerical framework that simultaneously offers modularity for rapid experimentation and discovery.
 
-The data in HCI is complicated and requires significant processing to enable the discovery and characterization of exoplanets. Angular differential imaging (ADI) is an observational technique for HCI that utilizes the Earth's rotation throughout a night of observing [@liu:2004; @marois_angular_2006]. Normally telescopes have optics to counter this rotation, but disabling these optics and taking images throughout the night will give a sequence of frames where the sky appears to rotate. The telescope optics produce a systematic noise that will not rotate with the sky, although it can slowly vary over time. The sequence of frames are pre-processed to calibrate the images and fix defects like bad pixels. The pre-processed images are then co-aligned and concatenated together into a data cube. ADI algorithms exploit the difference in rotation to approximate and subtract the systematics without removing too much signal from potential companions.
+Angular differential imaging (ADI) is an observational technique for HCI that utilizes the Earth's rotation throughout a night of observing [@liu:2004; @marois_angular_2006]. Normally telescopes have optics to counter this rotation, but disabling these optics and taking images throughout the night will give a sequence of frames where the sky appears to rotate. The telescope optics produce a systematic noise that will not rotate with the sky, although it can slowly vary over time. The sequence of frames are pre-processed to calibrate the images and fix defects like bad pixels. The pre-processed images are then co-aligned and concatenated together into a data cube. ADI algorithms exploit the difference in rotation to approximate and subtract the systematics without substantially overfitting the signals from potential companions.
 
-The difference in ADI algorithms is primarily in how the systematic signal is estimated. These algorithms can be applied in different ways, though- for example instead of estimating the systematics from the *target* star, a *reference* star can be used (reference differential imaging, RDI). Instead of using the entire frame, the cube can be processed in annuli corresponding to different circumstellar regions. These geometric techniques are independent of the underlying description of the algorithms. Similarly, GPU programming or out-of-core processing are computational techniques which appear like implementation details in comparison to the algorithms or how they are applied. Creating a *modular* and *generic* framework for HCI enables scientists to explore different algorithms and techniques flexibly, allowing more thorough and deeper investigations into the capabilities of HCI for finding exoplanets.
[…]
