+++ paper/paper.bib
@@ -226,6 +226,7 @@ archivePrefix = {ascl},
        eprint = {1506.001},
       journal = {Astrophysics Source Code Library},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2015ascl.soft06001W},
+          url = {https://ui.adsabs.harvard.edu/abs/2015ascl.soft06001W},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
