### 2020-12-15 #2843 @dfm
@giordano and @semaphoreP: Thanks for reviews! I really appreciate your time and your constructive suggestions.

@mileslucas: There are two things for you to do before we do the final processing:

1. I submitted a PR with some small edits to the manuscript. Take a look at that, merge it, and double check the manuscript to see if there are any other edits that you want to make. Then, comment `@whedon generate pdf` on this thread and double check the proofs. In particular, check that the author names and affiliations are correct. This is harder to change later!
2. Release `v0.5.3` and create an archive for the version using Zenodo or similar. **Note: the title and author list on that archive must _exactly_ match the manuscript**. Report the DOI of that archive to this thread.

Thanks all!!
