### 2020-12-16 #2843 @dfm
@mileslucas: Please update the metadata on the Zenodo archive to match the manuscript ([as requested above](https://github.com/openjournals/joss-reviews/issues/2843#issuecomment-745409496) in bold text :D). The title should be `ADI.jl: A Julia Package for High-Contrast Imaging` and the author list should be you and Michael Bottom. You don't need to make a new archive - you should be able to edit the metadata for the existing archive.
