### 2020-12-16 #2843 @mileslucas
@dfm I don't know what to say: every time I edit the DOI it says it has the correct information.
![Screen Shot 2020-12-16 at 2 33 01 PM](https://user-images.githubusercontent.com/14099459/102403259-9a8fb980-3fab-11eb-87d8-363643c1c4a8.png)

Saving doesn't change anything and when I publish I get a server error
