### 2020-12-16 #2843 @dfm
@mileslucas: Bizarre! It looks like you used some sort of GitHub actions workflow to publish your release? Perhaps it has something to do with that. Can you try uploading a new snapshot of the code with the right metadata to Zenodo directly? Thanks!
