### 2020-12-16 #2843 @mileslucas
Yeah, I'm using the zenodo feature that automatically updates with each github release. I sent a support request to zenodo to see if that can be resolved before I upload a new version. I'll ping here when I've got it figured out.
