### 2020-12-17 #2843 @kthyng
Everything looks to be in order! @mileslucas the summary in your paper is excellent! This is uncommon.

We can proceed.
